LEDs_serial/requirements.txt
getch
keyboard
mouse
numpy
pyaudio
pyserial
